third_party/blink/web_tests/editing/selection/dump-as-markup.html

<!DocType html>
<script src="../../resources/dump-as-markup.js"></script>

<div id="bar">foo<hr></div>
<input type=text>
<div contentEditable id="foo">
  foo   bar
</div>
<script>foo.firstChild.splitText(6);</script>
<span>foo</span><span>bar</span>
<div><div><b>sdf</b><i>sdf</i>sdf</div></div>
<iframe src='resources/dump-as-markup-iframe.html'></iframe>

<script>
Markup.description("Tests basic dump-as-markup functionality.");
window.getSelection().selectAllChildren(bar)
</script>
