third_party/blink/web_tests/editing/selection/dump-as-markup-expected.txt

Tests basic dump-as-markup functionality.
| <!DOCTYPE html>
| <html>
|   <head>
|     <script>
|       src="../../resources/dump-as-markup.js"
|     "

"
|   <body>
|     <div>
|       id="bar"
|       <#selection-anchor>
|       "foo"
|       <hr>
|       <#selection-focus>
|     "
"
|     <input>
|       type="text"
|       this.value=""
|       <shadow:root>
|         <div>
|     "
"
|     <div>
|       contenteditable=""
|       id="foo"
|       "
  foo"
|       "   bar
"
|     "
"
|     <script>
|       "foo.firstChild.splitText(6);"
|     "
"
|     <span>
|       "foo"
|     <span>
|       "bar"
|     "
"
|     <div>
|       <div>
|         <b>
|           "sdf"
|         <i>
|           "sdf"
|         "sdf"
|     "
"
|     <iframe>
|       src="resources/dump-as-markup-iframe.html"
|     "

"
|     <script>
|       "
Markup.description("Tests basic dump-as-markup functionality.");
window.getSelection().selectAllChildren(bar)
"
|     "
"

FRAME 0:
| <head>
|   "
  "
|   <title>
|     "Test"
|   "
  "
|   <style>
|     type="text/css"
|     "
  "
|   "
 "
| "
 "
| <body>
|   "
  "
|   <p>
|   "
 

"
